Add npm lockfile
This should help ensure that we're all using the same versions of all the dependencies (including sub-dependencies), which will avoid errors or possible minor differences in behaviour between our own locally-running versions and the instance deployed on Heroku.

The downside is that we all may need to run `npm install` more often.