Move the post publication date to the header (#709)
The date that a post was published in the design history is a pretty important bit of context for the whole post.  Whilst this is included prominently in the index pages, if someone arrives at a post via a direct link, they won't see the date straight away as it was at the end of the post.

This moves it to the header, after the title and before the main content. This follows a similar change on the [GOV.UK Design history](https://design-history.herokuapp.com/what-is-a-design-history/) project.