Adding support for tags on posts (#787)
This adds support for "tags" on posts, which then link to a tag page that lists all other posts (across all the services) the the same tag.

There's also a `/tags` page which lists all tags in alphabetical order, with the number of posts shown in brackets.

To get the tagging feature to work easily, the way that posts are grouped by service has been changed - these no longer get auto-tagged with a service name, and then grouped into a collection using that tag, but are instead grouped into a collection using the folder structure.

All the posts are also no longer tagged with `post`.

The tag pages aren’t currently paginated but instead show _all_ posts with that tag. We might need to paginate these in future.

We could also consider having "scoped" tag pages (eg `/apply-for-teacher-training/tagged/references`) in future, in addition to the cross-service tag pages.